refactor: couponClient 클래스 분리 및 인터페이스 추가
