+++ backend/src/main/java/com/woowacourse/thankoo/coupon/infrastructure/MeetingProvider.java
@@ -1,0 +1,8 @@
+package com.woowacourse.thankoo.coupon.infrastructure;
+
+import com.woowacourse.thankoo.coupon.infrastructure.integrate.dto.MeetingResponse;
+
+public interface MeetingProvider {
+
+    MeetingResponse getOnProgressMeeting(Long couponId);
+}
