+++ backend/src/main/java/com/woowacourse/thankoo/coupon/infrastructure/ReservationProvider.java
@@ -1,0 +1,8 @@
+package com.woowacourse.thankoo.coupon.infrastructure;
+
+import com.woowacourse.thankoo.coupon.infrastructure.integrate.dto.ReservationResponse;
+
+public interface ReservationProvider {
+
+    ReservationResponse getWaitingReservation(Long couponId);
+}
