+++ backend/src/main/java/com/woowacourse/thankoo/coupon/presentation/dto/CouponDetailResponse.java
@@ -1,8 +1,8 @@
 package com.woowacourse.thankoo.coupon.presentation.dto;
 
 import com.woowacourse.thankoo.coupon.domain.MemberCoupon;
-import com.woowacourse.thankoo.meeting.presentation.dto.MeetingResponse;
-import com.woowacourse.thankoo.reservation.presentation.dto.ReservationResponse;
+import com.woowacourse.thankoo.coupon.infrastructure.integrate.dto.MeetingResponse;
+import com.woowacourse.thankoo.coupon.infrastructure.integrate.dto.ReservationResponse;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
